GH-4629: Refactor tests to support mobile viewport.
