+++ projects/storefrontapp-e2e-cypress/cypress/integration/smoke/asm-mobile.e2e-spec.ts
@@ -1,0 +1,14 @@
+import * as asm from '../../helpers/asm';
+import { formats } from '../../sample-data/viewports';
+
+context(`${formats.mobile.width + 1}p resolution - ASM`, () => {
+  before(() => {
+    cy.window().then(win => win.sessionStorage.clear());
+  });
+
+  beforeEach(() => {
+    cy.viewport(formats.mobile.width, formats.mobile.height);
+  });
+
+  asm.asmTests();
+});
